Unzip and collect tracks into individual speaker directories
Now we have a separate directory for each speaker, we can start
to create the short sentences for transcription and later training.

The sentences will be created using pydub and silence spliting.

Because we are actually modifying data here, this will be put
into the processed folder. For now everything here is just the original
data but renamed and organized.