Add audio processing example using librosa
Using librosa we can use the split effect to get timestamps
when the audio is below a certain threshold

Although the current example works, I have to do this roundabout reconversion
to wav just so librosa can read in the file.

By looking at the documentation with ffmpeg-python, it seems that we can just
use the existing silencedetect filter from ffmpeg itself and avoid this
secondary library (which would be nice). We wouldn't need to create
an intermediate wav file and I think that the file would not completely
loading into memory using ffmpeg directly (not sure though)