Add ffmpeg-python silencedetect method
Not that great natively because it does not handle short sounds well or short pauses
Basically we have little control over how the sound is truncated compared to using silenceremove.

I want to look into using silenceremove now because it offered some pretty impressive results.
The only thing is though that it generates a combined sound file. I think that there is a 'timestamp'
option though and perhaps we can reprocess it to output at the particular rewritten timestamps.