Split audio with FFmpeg silencedetect and clean up directory
After much back and fourth, I decided to use the silencedetect filter
from FFmpeg because it avoids loading the entire audio file into memory,
it is fast (if non-accurate seeking is desired), it doesn't have to reencode,
and it does an okay job at splitting. Sometimes it does border on sentences
due to however the internal algorithm works but for our purposes it is enough for
vocoder training I believe.

I cleaned up the directory so that we have all the data processing modules in their
own folder, and the extract, split, and transcribe (currently empty) methods are
divided.

I also created a top-level 'main' script which will pass in the relevant parameters to
the submodules, instead of having every submodule take in the entire YAML config file.

For the 'accurate' parameter in the YAML file, we either copy without reencoding
(i.e. we go from .aac to .aac) but then we don't get accurate timestamp seeking.
Although I don't think the accuracy is all that big of a deal, and the speed-up
is quite significant.

Also add an ignore speaker in the processed directory in case we just want to
split sentences from a particular speaker and not remove existing directories

Next steps are to use Whisper to transcribe these split sentences.