Add code to generate chunks from ffmpeg audio file -> block processing
This uses FFmpeg's -ss and -t options to process fixed second
blocks of audio for use in subsequent processing. We can use
this in conjunction with the librosa splitting to avoid
handling the data all at once.

I also added a stride or 'hop' length so that we have overlap in
the samples. We will need to handle the processing of this
to make sure that we include any non-silent portions in the audio overlap