Add in scotts voice - works pretty well
its not perfect but at least i dont use a lot of RAM to do this
and i can experiment with the files. The goal is just to get
a lot of files and then use whisper and then see what the resulting
transcripts are. If the transcripts are not that great I will ditch
some of the files. I can use maybe some other chatbot or something
to identify if the files are good or not? Or maybe Whisper
will give me a score for each transcription and I'll know whether to
scrap it or not