Add initial code to run TTS pipeline from Discord to Tacotron2
Just a proof of concept but it works on my computers hardware. Takes
a while to actually output it but it does work. The text I type in
Discord is converted to a vocoder text.

Several interesting optimizations could be made:

could we output the audio sequentially? I'm not sure. For example, if
we asked to process a 32 word sentence can we stream out the audio
and have discord play out the words we have already?

Also to avoid saving to disk we want to stream it out io.buffered style