Clean up gitignore and add small dataset
The raw-small.zip contains just 3 nested zip files for 3 separate recordings.
there is enough audio to test sentence parsing and the overall TTS pipeline,
but not too much so it takes forever to process the files. Not to mention that
this small dataset is only 37.7 MB instead of the multiple GBs the full dataset is