Clean up and overhaul data extraction
Convert zipfile extraction into a recursive function and implement a simpler and cleaner
way to extract all nested zipfiles with a single funciton.

Now we don't have to rely too heavily on the format for the audio files ->
we will keep extracting nested zipfiles until we reach the bottom of a file
and we will extract the name and provide that to the group_tracks() function.

We return a speaker dict which contains information on the directories for the speakers and the
grouped up audio files