Overhaul directory structure and remove submodule (again)
Opt for a single directory with all the python files in one location.
It was getting too hectic putting these in separate directories since
there was a lot of overlap between some modules. One example is
that tacotron2 and the vocoder both need an audio dataset.

In the future, I might break this into sub-packages but now we will
stick with one uniform directory.

Also I am committing to NOT using an existing implementation of the
WaveRNN training loop. This is because of the bloat that some of these
modules have. I want to have an implementation which is similar to the tacotron2
training setup, where we are using the PyTorch implementations of the modules
and can use a fairly simple data preparation flow.

I flip-flopped back and fourth many times over this, and the main issue
that caused this undecisiveness was the test_wavernn.py file. It seems that the
GRU that's instantiated in the WaveRNN fails for somewhat 'long' sequence lengths.
I will follow what's in the pipeline_wavernn example repository from PyTorch to chunk
up these spectrograms into more manageable pieces, and that should resolve the issue.