Refactor dataset preparation and add collate_fns
Some major changes here. I ditched the existing dataset structure
to create a global "dataset" which just computes the mel spectrograms,
the audio files, and the text from the transcription mapping.

The advantage of this is now the tacotron2 and wavernn can consume
this 'global' dataset and manipulate it to their specific needs.
With PyTorch audio, tacotron2 and wavernn have different input requirements.
It is messy to try and fit both of these interfaces from a central dataset location.

Just from some initial training, loss is definitely going down. Still
have to sanity check and see what the output of the model is doing.

One important note is that before, we were hard-clipping the waveforms to
30 seconds. We need to do that when we load the audio file, since the transcriptions
were only valid up until that clipped value. Alternatively we could try fixing the
spectrogram length but it's better to fix the audio clip and thus take up less
memory anyway.

Also updated the stdout to give info on how many data samples we have seen and which
batch we are on