Verify that we can overfit to pure sine wave
Add in dummy_data parameter to just fit spectrogram prediction to a pure sine wave.
This was to make sure that the input/output flow of the tacotron2 was working as expected.
We are able to produce a pure sine wave.

I also stepped through the code and verified that the teacher forcing
resulted in spectrograms which were very close to the original spectrogram
during training.

However, at inference time, without the teacher forcing, the spectrograms look
pretty different.

Additional changes are allowing for more flexible specgram input shapes
when plotting, and capping the waveform when dataloading to the size which
was used to transcribe (by trimming to that length). Note that we still
need to increase the decoder length output potentially since it is by default 2000
but we expect training frame size to be >= 2401.